MM-19: Exchange warm welcomes
